# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tu Tiên Lộ Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /character:
    get:
      operationId: getCharacter
      tags: [character]
      summary: Get my character
      responses:
        "200":
          description: Character data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: No character yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createCharacter
      tags: [character]
      summary: Create character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
...
  /character/inspect/{characterId}:
    get:
      operationId: inspectCharacter
      tags: [character]
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: string
      summary: Inspect another player
      responses:
        "200":
          description: Character summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
...
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat messages
      parameters:
        - name: channel
          in: query
          required: false
          schema:
            type: string
            enum: [world, sect, private]
            default: world
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
  /chat/send:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatBody"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...
    CreateCharacterBody:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 20
      required:
        - name
    Character:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        realmKey:
          type: string
        realmName:
          type: string
        realmStage:
          type: integer
        title:
          type: string
          nullable: true
        exp:
          type: integer
        expRequired:
          type: integer
        level:
          type: integer
        hp:
          type: integer
        hpMax:
          type: integer
        mp:
          type: integer
        mpMax:
          type: integer
        stamina:
          type: integer
        staminaMax:
          type: integer
        power:
          type: integer
        spirit:
          type: integer
        speed:
          type: integer
        luck:
          type: integer
        daoVan:
          type: integer
        atk:
          type: integer
        def:
          type: integer
        crit:
          type: number
        linhThach:
          type: integer
        tienNgoc:
          type: integer
        congHien:
          type: integer
        primaryElement:
          type: string
          nullable: true
        spiritualRootGrade:
          type: string
          nullable: true
        cultivating:
          type: boolean
        sectId:
          type: string
          nullable: true
        sectName:
          type: string
          nullable: true
        pvpWins:
          type: integer
        pvpLosses:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - realmKey
        - realmName
        - realmStage
        - exp
        - expRequired
        - level
        - hp
        - hpMax
        - mp
        - mpMax
        - stamina
        - staminaMax
        - power
        - spirit
        - speed
        - luck
        - daoVan
        - atk
        - def
        - crit
        - linhThach
        - tienNgoc
        - congHien
        - cultivating
        - pvpWins
        - pvpLosses
        - createdAt
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        channel:
          type: string
        senderName:
          type: string
        senderRealm:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - channel
        - senderName
        - senderRealm
        - content
        - createdAt
    SendChatBody:
      type: object
      properties:
        channel:
          type: string
          enum: [world, sect, private]
          default: world
        content:
          type: string
          maxLength: 300
        targetId:
          type: string
          nullable: true
      required:
        - channel
        - content
...